refactor: DTO 분리
